Add files via upload
HTML code to customize the video size and position in the glasses. PHP code assists in simultaneously playing the three videos in three monitors after the video is customized in the glasses. PHP code is what should be used in experiments. HTML code can be used for demonstrations on a desktop.